added logic to create purchase and Stock

+++ src/main/java/com/manpower/pos/repository/PurchaseRepository.java
@@ -3,5 +3,5 @@ package com.manpower.pos.repository;
 import com.manpower.pos.model.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
+public interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
 }
